+++ python/cugraph/tests/test_betweenness_centrality.py
@@ -55,8 +55,8 @@ RESULT_DTYPE_OPTIONS = [np.float32, np.float64]
 # Comparison functions
 # =============================================================================
 def calc_betweenness_centrality(
-    graph_file,
-    directed=True,
+    Fixture_params_dts_sml,
+    directed=True, #no need of this
     k=None,
     normalized=False,
     weight=None,
@@ -65,7 +65,7 @@ def calc_betweenness_centrality(
     result_dtype=np.float64,
     use_k_full=False,
     multi_gpu_batch=False,
-    edgevals=False,
+    edgevals=False, #no need of this
 ):
     """ Generate both cugraph and networkx betweenness centrality
 
@@ -120,8 +120,9 @@ def calc_betweenness_centrality(
     G = None
     Gnx = None
 
-    G, Gnx = utils.build_cu_and_nx_graphs(graph_file, directed=directed,
-                                          edgevals=edgevals)
+    G, Gnx = Fixture_params_dts_sml
+    
+    
 
     assert G is not None and Gnx is not None
     if multi_gpu_batch:
@@ -297,48 +298,72 @@ def compare_scores(sorted_df, first_key, second_key, epsilon=DEFAULT_EPSILON):
 
 def prepare_test():
     gc.collect()
+    
+
+    
+# =============================================================================
+# Pytest Fixtures
+# =============================================================================    
+DIRECTED = [pytest.param(d) for d in DIRECTED_GRAPH_OPTIONS]
+DATASETS_SMALL = [pytest.param(d) for d in utils.DATASETS_SMALL]
+DATASETS_UNRENUMBERED = [pytest.param(d) for d in utils.DATASETS_UNRENUMBERED]
+WEIGHTED_GRAPH_OPTIONS = [pytest.param(w) for w in WEIGHTED_GRAPH_OPTIONS]
+
+
+fixture_params_dts_sml = utils.genFixtureParamsProduct(
+    (DATASETS_SMALL, "grph"),
+    (DIRECTED, "dirctd"),
+    (WEIGHTED_GRAPH_OPTIONS, "wgtd_gph_opts")) 
+
+fixture_params_dts_urnbrd = utils.genFixtureParamsProduct(
+    (DATASETS_UNRENUMBERED, "grph"),
+    (DIRECTED, "dirctd"),
+    (WEIGHTED_GRAPH_OPTIONS, "wgtd_gph_opts")) 
+
+@pytest.fixture(scope="module", params=fixture_params_dts_sml)
+def Fixture_params_dts_sml(request):
+    return utils.build_cu_and_nx_graphs(*request.param)
+
+
+
+
+@pytest.fixture(scope="module", params=fixture_params_dts_urnbrd)
+def Fixture_params_dts_urnbrd(request):
+    return utils.build_cu_and_nx_graphs(*request.param)
+
+
 
 
 # =============================================================================
 # Tests
 # =============================================================================
-@pytest.mark.parametrize("graph_file", utils.DATASETS_SMALL)
-@pytest.mark.parametrize("directed", DIRECTED_GRAPH_OPTIONS)
 @pytest.mark.parametrize("subset_size", SUBSET_SIZE_OPTIONS)
 @pytest.mark.parametrize("normalized", NORMALIZED_OPTIONS)
 @pytest.mark.parametrize("weight", [None])
 @pytest.mark.parametrize("endpoints", ENDPOINTS_OPTIONS)
 @pytest.mark.parametrize("subset_seed", SUBSET_SEED_OPTIONS)
 @pytest.mark.parametrize("result_dtype", RESULT_DTYPE_OPTIONS)
-@pytest.mark.parametrize("edgevals", WEIGHTED_GRAPH_OPTIONS)
 def test_betweenness_centrality(
-    graph_file,
-    directed,
+    Fixture_params_dts_sml,
     subset_size,
     normalized,
     weight,
     endpoints,
     subset_seed,
     result_dtype,
-    edgevals
 ):
     prepare_test()
     sorted_df = calc_betweenness_centrality(
-        graph_file,
-        directed=directed,
+        Fixture_params_dts_sml,
         normalized=normalized,
         k=subset_size,
         weight=weight,
         endpoints=endpoints,
         seed=subset_seed,
         result_dtype=result_dtype,
-        edgevals=edgevals,
     )
     compare_scores(sorted_df, first_key="cu_bc", second_key="ref_bc")
 
-
-@pytest.mark.parametrize("graph_file", utils.DATASETS_SMALL)
-@pytest.mark.parametrize("directed", DIRECTED_GRAPH_OPTIONS)
 @pytest.mark.parametrize("subset_size", [None])
 @pytest.mark.parametrize("normalized", NORMALIZED_OPTIONS)
 @pytest.mark.parametrize("weight", [None])
@@ -346,10 +371,8 @@ def test_betweenness_centrality(
 @pytest.mark.parametrize("subset_seed", SUBSET_SEED_OPTIONS)
 @pytest.mark.parametrize("result_dtype", RESULT_DTYPE_OPTIONS)
 @pytest.mark.parametrize("use_k_full", [True])
-@pytest.mark.parametrize("edgevals", WEIGHTED_GRAPH_OPTIONS)
 def test_betweenness_centrality_k_full(
-    graph_file,
-    directed,
+    Fixture_params_dts_sml,
     subset_size,
     normalized,
     weight,
@@ -357,14 +380,12 @@ def test_betweenness_centrality_k_full(
     subset_seed,
     result_dtype,
     use_k_full,
-    edgevals
 ):
     """Tests full betweenness centrality by using k = G.number_of_vertices()
     instead of k=None, checks that k scales properly"""
     prepare_test()
     sorted_df = calc_betweenness_centrality(
-        graph_file,
-        directed=directed,
+        Fixture_params_dts_sml,
         normalized=normalized,
         k=subset_size,
         weight=weight,
@@ -372,7 +393,6 @@ def test_betweenness_centrality_k_full(
         seed=subset_seed,
         result_dtype=result_dtype,
         use_k_full=use_k_full,
-        edgevals=edgevals
     )
     compare_scores(sorted_df, first_key="cu_bc", second_key="ref_bc")
 
@@ -381,25 +401,20 @@ def test_betweenness_centrality_k_full(
 #       the function operating the comparison inside is first proceeding
 #       to a random sampling over the number of vertices (thus direct offsets)
 #       in the graph structure instead of actual vertices identifiers
-@pytest.mark.parametrize("graph_file", utils.DATASETS_UNRENUMBERED)
-@pytest.mark.parametrize("directed", DIRECTED_GRAPH_OPTIONS)
 @pytest.mark.parametrize("subset_size", SUBSET_SIZE_OPTIONS)
 @pytest.mark.parametrize("normalized", NORMALIZED_OPTIONS)
 @pytest.mark.parametrize("weight", [None])
 @pytest.mark.parametrize("endpoints", ENDPOINTS_OPTIONS)
 @pytest.mark.parametrize("subset_seed", [None])
 @pytest.mark.parametrize("result_dtype", RESULT_DTYPE_OPTIONS)
-@pytest.mark.parametrize("edgevals", WEIGHTED_GRAPH_OPTIONS)
 def test_betweenness_centrality_fixed_sample(
-    graph_file,
-    directed,
+    Fixture_params_dts_urnbrd,
     subset_size,
     normalized,
     weight,
     endpoints,
     subset_seed,
     result_dtype,
-    edgevals
 ):
     """Test Betweenness Centrality using a subset
 
@@ -407,38 +422,31 @@ def test_betweenness_centrality_fixed_sample(
     """
     prepare_test()
     sorted_df = calc_betweenness_centrality(
-        graph_file,
-        directed=directed,
+        Fixture_params_dts_urnbrd,
         k=subset_size,
         normalized=normalized,
         weight=weight,
         endpoints=endpoints,
         seed=subset_seed,
         result_dtype=result_dtype,
-        edgevals=edgevals
     )
     compare_scores(sorted_df, first_key="cu_bc", second_key="ref_bc")
 
-
-@pytest.mark.parametrize("graph_file", utils.DATASETS_SMALL)
-@pytest.mark.parametrize("directed", DIRECTED_GRAPH_OPTIONS)
+    
 @pytest.mark.parametrize("subset_size", SUBSET_SIZE_OPTIONS)
 @pytest.mark.parametrize("normalized", NORMALIZED_OPTIONS)
 @pytest.mark.parametrize("weight", [[]])
 @pytest.mark.parametrize("endpoints", ENDPOINTS_OPTIONS)
 @pytest.mark.parametrize("subset_seed", SUBSET_SEED_OPTIONS)
 @pytest.mark.parametrize("result_dtype", RESULT_DTYPE_OPTIONS)
-@pytest.mark.parametrize("edgevals", WEIGHTED_GRAPH_OPTIONS)
 def test_betweenness_centrality_weight_except(
-    graph_file,
-    directed,
+    Fixture_params_dts_sml,
     subset_size,
     normalized,
     weight,
     endpoints,
     subset_seed,
     result_dtype,
-    edgevals
 ):
     """Calls betwenness_centrality with weight
 
@@ -448,53 +456,45 @@ def test_betweenness_centrality_weight_except(
     prepare_test()
     with pytest.raises(NotImplementedError):
         sorted_df = calc_betweenness_centrality(
-            graph_file,
-            directed=directed,
+            Fixture_params_dts_sml,
             k=subset_size,
             normalized=normalized,
             weight=weight,
             endpoints=endpoints,
             seed=subset_seed,
             result_dtype=result_dtype,
-            edgevals=edgevals
         )
         compare_scores(sorted_df, first_key="cu_bc", second_key="ref_bc")
 
 
-@pytest.mark.parametrize("graph_file", utils.DATASETS_SMALL)
-@pytest.mark.parametrize("directed", DIRECTED_GRAPH_OPTIONS)
+        
 @pytest.mark.parametrize("normalized", NORMALIZED_OPTIONS)
 @pytest.mark.parametrize("subset_size", SUBSET_SIZE_OPTIONS)
 @pytest.mark.parametrize("weight", [None])
 @pytest.mark.parametrize("endpoints", ENDPOINTS_OPTIONS)
 @pytest.mark.parametrize("subset_seed", SUBSET_SEED_OPTIONS)
 @pytest.mark.parametrize("result_dtype", [str])
-@pytest.mark.parametrize("edgevals", WEIGHTED_GRAPH_OPTIONS)
 def test_betweenness_invalid_dtype(
-    graph_file,
-    directed,
+    Fixture_params_dts_sml,
     subset_size,
     normalized,
     weight,
     endpoints,
     subset_seed,
     result_dtype,
-    edgevals
 ):
     """Test calls edge_betwenness_centrality an invalid type"""
 
     prepare_test()
     with pytest.raises(TypeError):
         sorted_df = calc_betweenness_centrality(
-            graph_file,
-            directed=directed,
+            Fixture_params_dts_sml,
             k=subset_size,
             normalized=normalized,
             weight=weight,
             endpoints=endpoints,
             seed=subset_seed,
             result_dtype=result_dtype,
-            edgevals=edgevals
         )
         compare_scores(sorted_df, first_key="cu_bc", second_key="ref_bc")
 
@@ -528,3 +528,5 @@ def test_betweenness_centrality_nx(
             print(f"{cugraph_bc[i][1]} and {cugraph_bc[i][1]}")
     print("Mismatches:", err)
     assert err < (0.01 * len(cugraph_bc))
+
+
